refactoring and enhancement commit

+++ src/main/java/com/nielvid/services/AccountService.java
@@ -7,10 +7,10 @@ import java.math.BigDecimal;
 
 public interface AccountService {
       public Account openAccount(User user, String accountType);
-      public void makeDeposit(int accountNumber, double amount);
-      public void withdrawFund(int accountNumber, double amount);
-      public void transferFund(int accountNumber, double amount);
-      public BigDecimal viewBalance(int accountNumber);
+      public void makeDeposit(long accountNumber, double amount);
+      public void withdrawFund(long accountNumber, double amount);
+      public void transferFund(long accountNumber, double amount);
+      public BigDecimal viewBalance(long accountNumber);
 //    public void viewTransactionHistory();
 //    public void closeAccount();
 }
